+++ Tests/HAPTests/EndpointTests.swift
@@ -38,7 +38,7 @@ class EndpointTests: XCTestCase {
         guard let accessory = jsonObject["accessories"]?.first else {
             return XCTFail("No accessory")
         }
-        XCTAssertEqual(accessory["aid"] as? UInt64, lamp.aid)
+        XCTAssertEqual(accessory["aid"] as? Int, lamp.aid)
         guard let services = accessory["services"] as? [[String: Any]] else {
             return XCTFail("No services")
         }
@@ -46,13 +46,13 @@ class EndpointTests: XCTestCase {
         guard let metaService = services.first(where: { ($0["type"] as? String) == "3E" }) else {
             return XCTFail("No meta")
         }
-        XCTAssertEqual(metaService["iid"] as? UInt64, 2)
+        XCTAssertEqual(metaService["iid"] as? Int, 2)
         XCTAssertEqual((metaService["characteristics"] as? [Any])?.count, 6)
 
         guard let lampService = services.first(where: { ($0["type"] as? String) == "43" }) else {
             return XCTFail("No lamp")
         }
-        XCTAssertEqual(lampService["iid"] as? UInt64, 9)
+        XCTAssertEqual(lampService["iid"] as? Int, 9)
 
         guard let lampCharacteristics = lampService["characteristics"] as? [[String: Any]] else {
             return XCTFail("No lamp characteristics")
@@ -73,12 +73,12 @@ class EndpointTests: XCTestCase {
             return XCTFail("No characteristics")
         }
 
-        guard let manufacturerCharacteristic = characteristics.first(where: { $0["iid"] as? UInt64 == lamp.info.manufacturer.iid }) else {
+        guard let manufacturerCharacteristic = characteristics.first(where: { $0["iid"] as? Int == lamp.info.manufacturer.iid }) else {
             return XCTFail("No manufacturer")
         }
         XCTAssertEqual(manufacturerCharacteristic["value"] as? String, "Bouke")
 
-        guard let nameCharacteristic = characteristics.first(where: { $0["iid"] as? UInt64 == lamp.info.name.iid }) else {
+        guard let nameCharacteristic = characteristics.first(where: { $0["iid"] as? Int == lamp.info.name.iid }) else {
             return XCTFail("No name")
         }
         XCTAssertEqual(nameCharacteristic["value"] as? String, "Night stand left")
@@ -488,15 +488,15 @@ class EndpointTests: XCTestCase {
                 return XCTFail("No characteristics")
             }
 
-            guard let light = characteristics.first(where: { $0["aid"] as? UInt64 == lightsensor.aid }) else {
+            guard let light = characteristics.first(where: { $0["aid"] as? Int == lightsensor.aid }) else {
                 return XCTFail("Could not get light aid")
             }
 
-            guard let therm = characteristics.first(where: { $0["aid"] as? UInt64 == thermostat.aid }) else {
+            guard let therm = characteristics.first(where: { $0["aid"] as? Int == thermostat.aid }) else {
                 return XCTFail("Could not get therm aid")
             }
 
-            guard let lampa = characteristics.first(where: { $0["aid"] as? UInt64 == lamp.aid }) else {
+            guard let lampa = characteristics.first(where: { $0["aid"] as? Int == lamp.aid }) else {
                 return XCTFail("Could not get lampa aid")
             }
 
@@ -525,8 +525,8 @@ class EndpointTests: XCTestCase {
             XCTAssertEqual(response.status, .multiStatus)
             let json = try! JSONSerialization.jsonObject(with: response.body!)  as! [String: [[String: Any]]]
             XCTAssertEqual(json["characteristics"]![0]["status"]! as! Int, HAPStatusCodes.writeOnly.rawValue)
-            XCTAssertEqual(json["characteristics"]![0]["aid"]! as! UInt64, aid)
-            XCTAssertEqual(json["characteristics"]![0]["iid"]! as! UInt64, iid)
+            XCTAssertEqual(json["characteristics"]![0]["aid"]! as! Int, aid)
+            XCTAssertEqual(json["characteristics"]![0]["iid"]! as! Int, iid)
         }
 
         // trying to read write only access and one with read access
